pyproject.toml
[project]
name = "small-tts"
version = "0.1.0"
description = "Small streaming TTS with Mimi codec"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "numpy>=1.24.0",
    "moshi>=0.2.0",
    "tqdm>=4.65.0",
    "tensorboard>=2.14.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]


